Added a few tests for weights loading
