+++ examples/src/main/kotlin/examples/cnn/fashionmnist/VGGMnist.kt
@@ -3,13 +3,11 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
  */
 
-package examples.keras.cifar10
+package examples.cnn.fashionmnist
 
 import org.jetbrains.kotlinx.dl.api.core.Sequential
-import org.jetbrains.kotlinx.dl.api.core.WritingMode
 import org.jetbrains.kotlinx.dl.api.core.activation.Activations
 import org.jetbrains.kotlinx.dl.api.core.initializer.HeNormal
-import org.jetbrains.kotlinx.dl.api.core.initializer.HeUniform
 import org.jetbrains.kotlinx.dl.api.core.layer.Dense
 import org.jetbrains.kotlinx.dl.api.core.layer.Flatten
 import org.jetbrains.kotlinx.dl.api.core.layer.Input
@@ -20,21 +18,21 @@ import org.jetbrains.kotlinx.dl.api.core.loss.Losses
 import org.jetbrains.kotlinx.dl.api.core.metric.Metrics
 import org.jetbrains.kotlinx.dl.api.core.optimizer.Adam
 import org.jetbrains.kotlinx.dl.datasets.Dataset
-import java.io.File
+import org.jetbrains.kotlinx.dl.datasets.handlers.*
 
-private const val PATH_TO_MODEL = "savedmodels/vgg11"
-private const val EPOCHS = 20
-private const val TRAINING_BATCH_SIZE = 500
+private const val EPOCHS = 5
+private const val TRAINING_BATCH_SIZE = 100
 private const val TEST_BATCH_SIZE = 1000
 private const val NUM_LABELS = 10
-private const val NUM_CHANNELS = 3L
-private const val IMAGE_SIZE = 32L
-private const val TRAIN_TEST_SPLIT_RATIO = 0.8
+private const val NUM_CHANNELS = 1L
+private const val IMAGE_SIZE = 28L
+private const val SEED = 12L
 
-/** This model is an implementation of VGG'11 model with reduced number of neurons in each layer (2 times for Conv2D and 8 times for Dense).
- *
- * @see <a href="https://arxiv.org/abs/1409.1556">
- *     Very Deep Convolutional Networks for Large-Scale Image Recognition:[Karen Simonyan, Andrew Zisserman, 2015]</a>
+val heNormal = HeNormal(SEED)
+
+/**
+ * Kotlin implementation of VGG'11 on Keras.
+ * Architecture could be copied here: https://medium.com/@amir_hf8/implementing-vgg13-for-mnist-dataset-in-tensorflow-abc1460e2b93
  */
 private val vgg11 = Sequential.of(
     Input(
@@ -47,8 +45,8 @@ private val vgg11 = Sequential.of(
         kernelSize = longArrayOf(3, 3),
         strides = longArrayOf(1, 1, 1, 1),
         activation = Activations.Relu,
-        kernelInitializer = HeNormal(),
-        biasInitializer = HeNormal(),
+        kernelInitializer = heNormal,
+        biasInitializer = heNormal,
         padding = ConvPadding.SAME
     ),
     MaxPool2D(
@@ -60,8 +58,8 @@ private val vgg11 = Sequential.of(
         kernelSize = longArrayOf(3, 3),
         strides = longArrayOf(1, 1, 1, 1),
         activation = Activations.Relu,
-        kernelInitializer = HeNormal(),
-        biasInitializer = HeNormal(),
+        kernelInitializer = heNormal,
+        biasInitializer = heNormal,
         padding = ConvPadding.SAME
     ),
     MaxPool2D(
@@ -73,8 +71,8 @@ private val vgg11 = Sequential.of(
         kernelSize = longArrayOf(3, 3),
         strides = longArrayOf(1, 1, 1, 1),
         activation = Activations.Relu,
-        kernelInitializer = HeNormal(),
-        biasInitializer = HeNormal(),
+        kernelInitializer = heNormal,
+        biasInitializer = heNormal,
         padding = ConvPadding.SAME
     ),
     Conv2D(
@@ -82,8 +80,8 @@ private val vgg11 = Sequential.of(
         kernelSize = longArrayOf(3, 3),
         strides = longArrayOf(1, 1, 1, 1),
         activation = Activations.Relu,
-        kernelInitializer = HeNormal(),
-        biasInitializer = HeNormal(),
+        kernelInitializer = heNormal,
+        biasInitializer = heNormal,
         padding = ConvPadding.SAME
     ),
     MaxPool2D(
@@ -95,8 +93,8 @@ private val vgg11 = Sequential.of(
         kernelSize = longArrayOf(3, 3),
         strides = longArrayOf(1, 1, 1, 1),
         activation = Activations.Relu,
-        kernelInitializer = HeNormal(),
-        biasInitializer = HeNormal(),
+        kernelInitializer = heNormal,
+        biasInitializer = heNormal,
         padding = ConvPadding.SAME
     ),
     Conv2D(
@@ -104,8 +102,8 @@ private val vgg11 = Sequential.of(
         kernelSize = longArrayOf(3, 3),
         strides = longArrayOf(1, 1, 1, 1),
         activation = Activations.Relu,
-        kernelInitializer = HeNormal(),
-        biasInitializer = HeNormal(),
+        kernelInitializer = heNormal,
+        biasInitializer = heNormal,
         padding = ConvPadding.SAME
     ),
     MaxPool2D(
@@ -113,21 +111,21 @@ private val vgg11 = Sequential.of(
         strides = intArrayOf(1, 2, 2, 1)
     ),
     Conv2D(
-        filters = 256,
+        filters = 128,
         kernelSize = longArrayOf(3, 3),
         strides = longArrayOf(1, 1, 1, 1),
-        activation = Activations.Elu,
-        kernelInitializer = HeNormal(),
-        biasInitializer = HeNormal(),
+        activation = Activations.Relu,
+        kernelInitializer = heNormal,
+        biasInitializer = heNormal,
         padding = ConvPadding.SAME
     ),
     Conv2D(
-        filters = 256,
+        filters = 128,
         kernelSize = longArrayOf(3, 3),
         strides = longArrayOf(1, 1, 1, 1),
-        activation = Activations.Elu,
-        kernelInitializer = HeNormal(),
-        biasInitializer = HeNormal(),
+        activation = Activations.Relu,
+        kernelInitializer = heNormal,
+        biasInitializer = heNormal,
         padding = ConvPadding.SAME
     ),
     MaxPool2D(
@@ -136,35 +134,35 @@ private val vgg11 = Sequential.of(
     ),
     Flatten(),
     Dense(
-        outputSize = 512,
+        outputSize = 2048,
         activation = Activations.Relu,
-        kernelInitializer = HeNormal(),
-        biasInitializer = HeUniform()
+        kernelInitializer = heNormal,
+        biasInitializer = heNormal
     ),
     Dense(
-        outputSize = 512,
+        outputSize = 1000,
         activation = Activations.Relu,
-        kernelInitializer = HeNormal(),
-        biasInitializer = HeUniform()
+        kernelInitializer = heNormal,
+        biasInitializer = heNormal
     ),
     Dense(
-        outputSize = NUM_LABELS,
+        outputSize = NUM_LABELS, // changed due to 10 classes instead of
         activation = Activations.Linear,
-        kernelInitializer = HeNormal(),
-        biasInitializer = HeUniform()
+        kernelInitializer = heNormal,
+        biasInitializer = heNormal
     )
 )
 
 fun main() {
-    val dataset = Dataset.create(
-        IMAGES_ARCHIVE,
-        LABELS_ARCHIVE,
-        NUM_LABELS,
-        ::extractCifar10Images,
-        ::extractCifar10Labels
+    val (train, test) = Dataset.createTrainAndTestDatasets(
+        FASHION_TRAIN_IMAGES_ARCHIVE,
+        FASHION_TRAIN_LABELS_ARCHIVE,
+        FASHION_TEST_IMAGES_ARCHIVE,
+        FASHION_TEST_LABELS_ARCHIVE,
+        NUMBER_OF_CLASSES,
+        ::extractFashionImages,
+        ::extractFashionLabels
     )
-
-    val (train, test) = dataset.split(TRAIN_TEST_SPLIT_RATIO)
 
     vgg11.use {
         it.compile(
@@ -173,11 +171,7 @@ fun main() {
             metric = Metrics.ACCURACY
         )
 
-        it.summary()
-
         it.fit(dataset = train, epochs = EPOCHS, batchSize = TRAINING_BATCH_SIZE, verbose = true)
-
-        it.save(File(PATH_TO_MODEL), writingMode = WritingMode.OVERRIDE)
 
         val accuracy = it.evaluate(dataset = test, batchSize = TEST_BATCH_SIZE).metrics[Metrics.ACCURACY]
