+++ examples/src/main/kotlin/examples/cnn/mnist/LeNetClassicWithAlternativeLosses.kt
@@ -3,10 +3,11 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
  */
 
-package examples.keras.mnist
+package examples.cnn.mnist
 
 import org.jetbrains.kotlinx.dl.api.core.Sequential
 import org.jetbrains.kotlinx.dl.api.core.activation.Activations
+import org.jetbrains.kotlinx.dl.api.core.history.EpochTrainingEvent
 import org.jetbrains.kotlinx.dl.api.core.initializer.Constant
 import org.jetbrains.kotlinx.dl.api.core.initializer.GlorotNormal
 import org.jetbrains.kotlinx.dl.api.core.initializer.Zeros
@@ -24,7 +25,7 @@ import org.jetbrains.kotlinx.dl.datasets.Dataset
 import org.jetbrains.kotlinx.dl.datasets.handlers.*
 
 private const val EPOCHS = 3
-private const val TRAINING_BATCH_SIZE = 1000
+private const val TRAINING_BATCH_SIZE = 2000
 private const val NUM_CHANNELS = 1L
 private const val IMAGE_SIZE = 28L
 private const val SEED = 12L
@@ -79,7 +80,7 @@ private val lenet5Classic = Sequential.of(
     ),
     Dense(
         outputSize = NUMBER_OF_CLASSES,
-        activation = Activations.Linear,
+        activation = Activations.Sigmoid,
         kernelInitializer = GlorotNormal(SEED),
         biasInitializer = Constant(0.1f)
     )
@@ -96,19 +97,31 @@ fun main() {
         ::extractLabels
     )
 
+    val (newTrain, validation) = train.split(0.95)
+
     lenet5Classic.use {
         it.compile(
-            optimizer = Adam(clipGradient = ClipGradientByValue(0.1f)),
-            loss = Losses.SOFT_MAX_CROSS_ENTROPY_WITH_LOGITS,
+            optimizer = Adam(clipGradient = ClipGradientByValue(0.5f)),
+            loss = Losses.RMSE,
             metric = Metrics.ACCURACY
         )
 
         it.summary()
 
-        it.fit(dataset = train, epochs = EPOCHS, batchSize = TRAINING_BATCH_SIZE, verbose = true)
+        val history = it.fit(
+            trainingDataset = newTrain,
+            validationDataset = validation,
+            validationBatchSize = TRAINING_BATCH_SIZE,
+            epochs = EPOCHS,
+            trainBatchSize = TRAINING_BATCH_SIZE,
+            verbose = true
+        )
 
         val accuracy = it.evaluate(dataset = test, batchSize = TEST_BATCH_SIZE).metrics[Metrics.ACCURACY]
 
         println("Accuracy: $accuracy")
+
+        val accuracyByEpoch = history[EpochTrainingEvent::metricValue]
+        println(accuracyByEpoch.contentToString())
     }
 }
