--- examples/src/main/kotlin/examples/inference/keras/mashaasked/loadConfig.kt
@@ -1,31 +1,0 @@
-/*
- * Copyright 2020 JetBrains s.r.o. and Kotlin Deep Learning project contributors. All Rights Reserved.
- * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
- */
-
-package examples.inference.keras.mashaasked
-
-import io.jhdf.HdfFile
-import org.jetbrains.kotlinx.dl.api.core.Sequential
-import org.jetbrains.kotlinx.dl.api.core.loss.Losses
-import org.jetbrains.kotlinx.dl.api.core.metric.Metrics
-import org.jetbrains.kotlinx.dl.api.core.optimizer.Adam
-import org.jetbrains.kotlinx.dl.api.inference.keras.loadWeights
-import org.jetbrains.kotlinx.dl.api.inference.keras.recursivePrintGroupInHDF5File
-import java.io.File
-
-fun main() {
-    val JSONConfig = File("C:\\modelConfig.json")
-    val model = Sequential.loadModelConfiguration(JSONConfig)
-    val modelDirectory = HdfFile(File("C:\\weights"))
-
-    recursivePrintGroupInHDF5File(modelDirectory, modelDirectory)
-    model.compile(
-        optimizer = Adam(),
-        loss = Losses.SOFT_MAX_CROSS_ENTROPY_WITH_LOGITS,
-        metric = Metrics.ACCURACY
-    )
-
-    model.summary()
-    model.loadWeights(modelDirectory)
-}
