+++ examples/src/main/kotlin/examples/inference/savedmodel/Print_out_metadata.kt
@@ -3,7 +3,7 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
  */
 
-package examples.inference
+package examples.inference.savedmodel
 
 import org.jetbrains.kotlinx.dl.api.inference.savedmodel.SavedModel
