+++ examples/src/main/kotlin/examples/inference/savedmodel/Simple_Mnist_inference_with_tensor_names.kt
@@ -3,12 +3,16 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE.txt file.
  */
 
-package examples.inference
+package examples.inference.savedmodel
 
+import org.jetbrains.kotlinx.dl.api.core.metric.Metrics
+import org.jetbrains.kotlinx.dl.api.inference.savedmodel.Input
+import org.jetbrains.kotlinx.dl.api.inference.savedmodel.Output
+import org.jetbrains.kotlinx.dl.api.inference.savedmodel.SavedModel
 import org.jetbrains.kotlinx.dl.datasets.Dataset
 import org.jetbrains.kotlinx.dl.datasets.handlers.*
 
-private const val PATH_TO_MODEL = "api/src/main/resources/savedmodel"
+private const val PATH_TO_MODEL = "api/src/main/resources/models/savedmodel"
 
 fun main() {
     val (train, test) = Dataset.createTrainAndTestDatasets(
@@ -21,23 +25,21 @@ fun main() {
         ::extractLabels
     )
 
-    /*val mnistModel = prepareModelForInference {
-        loadModel(PATH_TO_MODEL)
-        reshape(::reshapeInput)
-        input(Input.PLACEHOLDER)
-        output(Output.ARGMAX)
-    }
+    SavedModel.load(PATH_TO_MODEL).use {
+        println(it)
 
-    mnistModel.use {
-        val prediction = it.predict(train.getX(0))
+        it.reshape(::reshapeInput)
+        it.input(Input.PLACEHOLDER)
+        it.output(Output.ARGMAX)
+
+        val prediction = it.predict(train.getX(0), "Placeholder", "ArgMax")
 
         println("Predicted Label is: $prediction")
         println("Correct Label is: " + train.getLabel(0))
 
-        val predictions = it.predictAll(test)
+        val predictions = it.predictAll(test, "Placeholder", "ArgMax")
         println(predictions.toString())
 
         println("Accuracy is : ${it.evaluate(test, Metrics.ACCURACY)}")
-    }*/
+    }
 }
-
